+++ R/outlier.R
@@ -65,7 +65,7 @@ outlie <- function(data,UERE=10,standardize=FALSE,plot=TRUE,...)
 
 #########################
 # plot outlier information with error bars
-plot.outlie <- function(x,level=0.95,...)
+plot.outlie <- function(x,level=0.95,units=TRUE,...)
 {
   n <- nrow(x)
   v <- x$speed
@@ -74,12 +74,12 @@ plot.outlie <- function(x,level=0.95,...)
   v <- sapply(1:n,function(i){tnorm.hdr(v[i],x$VAR.speed[i],level=level)})
   d <- sapply(1:n,function(i){tnorm.hdr(d[i],x$VAR.distance[i],level=level)})
 
-  # unit conversions ...
-  UNITS <- unit(v,dimension='speed',concise=TRUE)
+  # unit conversions ... (simpler)
+  UNITS <- unit(v,dimension='length',concise=TRUE,SI=!units)
   v <- v/UNITS$scale
-  ylab <- paste0("Minimum speed (",UNITS$name,")")
+  ylab <- paste0("Minimum speed (",UNITS$name,"/s)")
 
-  UNITS <- unit(d,dimension='length',concise=TRUE)
+  UNITS <- unit(d,dimension='length',concise=TRUE,SI=!units)
   d <- d/UNITS$scale
   xlab <- paste0("Core deviation (",UNITS$name,")")